+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## Changed
+- use `dataTimestamp` as part of consensus
+- use `chain.getStatus` to pull all consensus data
+
+## Removed
+- `ValidatorRegistry`
 
 ## [0.10.6] - 2021-04-15